Revert " Reset chatbot context & implemented image search through google-image module."
