+++ config/api.js
@@ -1,5 +1,3 @@
 module.exports = {
-    'apiai': process.env.API || process.env.LOCAL_APIAI,
-    'gi-client-id' : process.env.GI_CLIENT_ID,
-    'gi-api-key' : process.env.GI_API_KEY
+    'apiai': process.env.API || process.env.LOCAL_APIAI
 }
