+++ package.json
@@ -23,7 +23,6 @@
     "chai": "^3.5.0",
     "classnames": "^2.2.5",
     "express": "^4.14.1",
-    "google-images": "^2.1.0",
     "mocha": "^3.2.0",
     "mongoose": "^4.8.4",
     "node-sass": "^4.5.0",
